restructuring GA parts
